+++ src/overtone/studio/mixer.clj
@@ -241,20 +241,20 @@
 (defn add-instrument
   "Add an instrument to the session."
   [inst]
-  (let [i-name (:name inst)]
-    (swap! studio* update-in [:instruments i-name] (fn [_] inst))
+  (let [i-name (:full-name inst)]
+    (swap! studio* assoc-in [:instruments i-name] inst)
     i-name))
 
 (defn remove-instrument
   "Remove an instrument from the session."
-  [i-name]
-  (swap! studio* (fn [s]
-                   (let [insts (:instruments s)
-                         insts (dissoc insts (name i-name))]
-                     (assoc s :instruments insts))))
-  (event :inst-removed :inst-name i-name))
+  [full-name]
+  (swap! studio* update :instruments dissoc full-name)
+  (event :inst-removed :inst-name full-name))
 
 (defn clear-instruments
   "Clear all instruments from the session."
   []
-  (swap! studio* assoc :instruments {}))
+  (let [[{:keys [instruments]} _]
+        (swap-vals! studio* assoc :instruments {})]
+    (doseq [[_name inst] instruments]
+      (group-free (:group inst)))))