+++ src/overtone/studio/pattern.clj
@@ -39,7 +39,12 @@
   "Like next, but with \"pattern semantics\". Lazily flattens nested sequences.
 
   When called with `dur` argument, will honor the `:dur` in metadata, possibly
-  repeating elements if their duration hasn't finished."
+  repeating elements if their duration hasn't finished.
+
+      (pbind {:degree [:i :i :iii :ii :i :iv :v]
+              :dur    [1/2 1 1/2 1/2 1/2 1/2 1/2]
+              :root   ^{:dur 8} [:c3 :d3 :b2 :c3]})
+  "
   ([s]
    (when (sequential? s)
      (let [[x & xs] s]