+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Unreleased
 
+## Added
+
+## Fixed
+
+## Changed
+
+# 0.15.3295 (2024-10-24 / d354b4f)
+
 This release features a significant update of the Pattern Library introduced in
 0.14. These changes make it more suitable for live programming, improving the
 behavior when patterns which are currently playing are being redefined. The
@@ -44,6 +52,9 @@ reconcile the two.
 - [doc] Explain how to get `mda-piano` from sc3-plugins
 - [doc] Add a tip about scsynth and Homebrew
 - [implementation] Warn when imported functions into `overtone.live` conflict
+- [studio] The instrument mixers now contains safety precautions: limiter+check-bad-values
+- [studio] Allow passing additional arguments to `inst-fx!`
+- [studio] Add "aux send" API (`aux-bus`, `aux-ctl`)
 
 ## Fixed
 
@@ -74,6 +85,8 @@ reconcile the two.
 ## Changed
 
 - [at-at] Bumped at-at to 1.4.65, which fixes reflection warnings.
+- [insts] sampled-piano now takes an `:amp` arg, as per synth/inst conventions
+- [studio] `clear-instruments` now also frees the instruments on the server
 
 # 0.14.3199 (2024-05-19 / 5d1c1ed)